control_timeshift: change timeshift from int to float

+++ gstlal-ugly/gst/control/control_timeshift.h
@@ -54,7 +54,7 @@ typedef struct {
 typedef struct {
 	GstBaseTransform element;
 
-	int shift;
+	float shift;
 	gint rate; // for offset
 
 } ControlTimeshift;
